Added classes dir Added DeviceTests
