+++ composer.json
@@ -6,7 +6,8 @@
   },
   "autoload": {
     "psr-4": {
-      "components\\": "components/"
+      "components\\": "components/",
+      "classes\\": "classes/"
     }
   }
 }
